Add initial bare support
Add a new "bare" attribute for projects. A bare project does not have
an associated revision, and the .git directory contents are instead
stored in the directory, as would be the case for a bare clone. (In
fact, they're usually going to be symlinks into the copy in the .repo
subdirectory.)

This is not fully implemented, but works for "repo init" and "repo sync".

Signed-off-by: Peter Seebach <peter.seebach@windriver.com>

Fixed a few minor issues in the original version related to directory
name and spacing.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>