Additional changes to implement the bare clone functionality
In a bare repo, ignore any revisionIds specified.  They have no meaning, but
if specified in the <remote /> tag there is no way to turn it off.

In all cases, avoid calling _Checkout.  bare clones should not have a checkout
performed on them.  Also make any calls to _Checkout raise an exception so if
any additional cases are found we will know about it.

When updating the repo, we want to always allow the the head to be updated,
because it's not actually a working directory.

All repositories are required to have a HEAD file.  We simply follow the
behavior from another part of repo and write one in the bare case.  It may
refer to a branch that does not exist.  This is OK, as long as the format of
the file is valid!

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>