progress.py: emit progress when stderr is not tty
Issue: LINCD-845

when wrlinux setup run repo sync as subprogress,
we make stderr/stdout as subprocess.PIPE, which is not
tty, so the progress info not emited.

fix by also emit progress when stderr is not tty, so
below info can be transfered to parent progess.

Fetching projects: 100% (317/317), done.
Checking out projects: 100% (317/317), done.

Signed-off-by: Changqing Li <changqing.li@windriver.com>